+++ Danzer D3D12 Engine/Code/Engine/Rendering/Models/ModelData.h
@@ -81,7 +81,7 @@ public:
 		m_transformBuffer.UpdateBuffer(reinterpret_cast<UINT8*>(&m_instanceTransforms[0]), (UINT)m_instanceTransforms.size(), frameIndex); 
 	}
 	void UpdateTransformInstanceBuffer(std::vector<Mat4f>& transform, const UINT frameIndex) {
-		m_transformBuffer.UpdateBuffer(reinterpret_cast<UINT8*>(&transform[0]), transform.size(), frameIndex);
+		m_transformBuffer.UpdateBuffer(reinterpret_cast<UINT8*>(&transform[0]), (UINT)transform.size(), frameIndex);
 	}
 	void UpdatedMaterialBuffer(UINT frameIndex) {
 		//MaterialBuffer::Data data;
